API Development

+++ backend/controllers/postController.js
@@ -1,11 +1,20 @@
 const asyncHandler = require("express-async-handler");
-const Post  = require('../models/postModel')
+const Post = require("../models/postModel");
+const User = require("../models/userModel");
 
 //@desc Get posts
 //@route GET/api/posts
 //@access private
-const getPosts = asyncHandler(async (req, res) => {
-  const posts = await Post.find()
+const getAllPosts = asyncHandler(async (req, res) => {
+  const posts = await Post.find();
+  res.status(200).json(posts);
+});
+
+//@desc Get posts
+//@route GET/api/posts
+//@access private
+const getUserPosts = asyncHandler(async (req, res) => {
+  const posts = await Post.find({ user: req.user.id });
   res.status(200).json(posts);
 });
 
@@ -17,13 +26,14 @@ const createPost = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Please add a text");
   }
-  
-  const post = await Post.create({
-    title: req.body.title,
-    description : req.body.description,
-    image: req.body.image
 
-  })
+  const post = await Post.create({
+    user: req.user.id,
+    author: `${req.user.firstName} ${req.user.lastName}`,
+    title: req.body.title,
+    description: req.body.description,
+    image: req.body.image,
+  });
 
   res.status(200).json(post);
 });
@@ -32,16 +42,30 @@ const createPost = asyncHandler(async (req, res) => {
 //@route PUT /api/posts/ceate
 //@access private
 const updatePost = asyncHandler(async (req, res) => {
+  const post = await Post.findById(req.params.id);
 
-  const post = await Post.findById(req.params.id)
-
-  if(!post){
-    res.status(400)
-    throw new Error('Post Not Found')
+  if (!post) {
+    res.status(400);
+    throw new Error("Post Not Found");
   }
 
-  const updatedPost = await Post.findByIdAndUpdate(req.params.id, req.body, {new:true})
+  const user = await User.findById(req.user.id);
 
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+
+  //make sure logged user match
+  if (post.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user not authorized");
+  }
+
+  const updatedPost = await Post.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+  });
 
   res.status(200).json(updatedPost);
 });
@@ -50,20 +74,34 @@ const updatePost = asyncHandler(async (req, res) => {
 //@route PUT /api/posts/ceate
 //@access private
 const deletePost = asyncHandler(async (req, res) => {
+  const post = await Post.findById(req.params.id);
 
-  const post = await Post.findById(req.params.id)
-
-  if(!post){
-    res.status(400)
-    throw new Error('Post Not Found')
+  if (!post) {
+    res.status(400);
+    throw new Error("Post Not Found");
   }
 
-  await Post.deleteOne()
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+
+  //make sure logged user match
+  if (post.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user not authorized");
+  }
+
+  await Post.deleteOne();
   res.status(200).json({ id: req.params.id });
 });
 
 module.exports = {
-  getPosts,
+  getAllPosts,
+  getUserPosts,
   createPost,
   updatePost,
   deletePost,
